Refactor Search.js to Redux and update Header.js.
